无用module已经移除，替换新的SDK已经跑起，开始做歌词等功能开发

+++ editor-core/src/main/java/com/ss/ugc/android/editor/core/utils/FileUtil.java
@@ -450,8 +450,8 @@ public class FileUtil {
      * @param timeMs
      * @return
      */
-    public static String stringForTime(int timeMs){
-        int totalSeconds = timeMs/1000;
+    public static String stringForTime(Long timeMs){
+        int totalSeconds = (int) (timeMs/1000);
         int seconds = totalSeconds % 60;
         int minutes = (totalSeconds/60)%60;
         int hours = totalSeconds/3600;
